lawn-9p: fix unnecessary cast warnings
In some cases, we need to allow unnecessary casts because not all
systems will have a 64-bit device or inode number or size.  In one case,
we have a truly unneeded cast to u8 which can be removed.

+++ lawn-9p/src/server/implementation.rs
@@ -711,7 +711,7 @@ impl<
                 trace!(
                     self.logger,
                     "9P: unknown message {:02x} size {:08x} tag {:02x}{:02x}",
-                    buf[4] as u8,
+                    buf[4],
                     size,
                     tag.0[0],
                     tag.0[1]
